refactor: 클래스 이름 변경  RevenueCalculator.js -> LottoRevenueCalculator.js
