+++ src/classes/LottoRevenueCalculator.js
